requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
Django==1.9.6
django-rest-auth==0.7.0
djangorestframework==3.3.3
factory-boy==2.7.0
fake-factory==0.5.7
gnureadline==6.3.3
httpie==0.9.3
ipython==4.2.0
ipython-genutils==0.1.0
pexpect==4.1.0
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
